añadiendo un nuevo componente card
