+++ src/components/ProductCard.js
@@ -1,7 +1,6 @@
 import { useState, useEffect } from 'react';
 import axios from 'axios';
-import Contador from './Contador';
-import { Link } from 'react-router-dom';
+import Card from './Card';
 
 const ProductCard = () => {
   const [card, setCard] = useState([]);
@@ -15,35 +14,12 @@ const ProductCard = () => {
     getCard();
   }, []);
 
-  const generateTime = ((min, max) => {
-    return Math.floor(Math.random() * (max - min) + min);
-  })(120000, 240000);
-  const now_ml = new Date().getTime();
-
-  const targetTime = now_ml + generateTime;
-
   return (
     <>
       {card.length === 0 ? (
         <h3>Cargando...</h3>
       ) : (
-        card.map((el) => {
-          return (
-            <div className="card" key={el.id}>
-              <img src={el.image} alt="Product" />
-              <div className="card__content">
-                <h4 className="card__title">{el.title}</h4>
-                <div className="card__buttons">
-                  {/* <Contador generateTime={generateTime(120000, 240000)} /> */}
-                  <Contador targetTime={targetTime} />
-                  <Link to={`/detail/${el.id}`} className="card__btn-link">
-                    Ir a detalles
-                  </Link>
-                </div>
-              </div>
-            </div>
-          );
-        })
+        card.map((el, index) => <Card key={index} el={el} />)
       )}
     </>
   );
